+++ src/main/java/com/codecool/budapestgo/dao/repository/PassRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface PassRepository extends JpaRepository<Pass,Integer> {
+public interface PassRepository extends JpaRepository<Pass, Long> {
     @Query("SELECT p FROM Pass p WHERE p.expireTime < NOW() AND  p.client.id = :id")
-    List<Pass> getAllExpiredPassesByClient_id(Integer id);
+    List<Pass> getAllExpiredPassesByClient_id(Long id);
     @Query("SELECT p FROM Pass p WHERE p.expireTime > NOW() AND  p.client.id = :id")
-    List<Pass> getActivePassesByClient_id(Integer id);
+    List<Pass> getActivePassesByClient_id(Long id);
 }
