+++ src/main/java/com/codecool/budapestgo/dao/repository/StopRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StopRepository extends JpaRepository<Stop,Integer> {
+public interface StopRepository extends JpaRepository<Stop, Long> {
     Optional<Stop> getStopByName(@NonNull String name);
 }
